adjusted accordion details to interact with modules

+++ resources/views/livewire/course/course-table.blade.php
@@ -1,9 +1,9 @@
 <div>
-    <h1 class="my-4 text-xl font-bold uppercase">Created Courses</h1>
+    <h1 class="my-4 text-xl font-bold uppercase">{{ $title }}</h1>
     <hr class="mb-4">
 
 
-    <div class="flex flex-wrap justify-center ">
+    <div class="flex flex-wrap items-start justify-center ">
         @forelse ($courses as $course)
             @php
                 $backgroundUrl = asset('storage/' . $course->course_cover_photo);
@@ -24,22 +24,24 @@
                         </div>
 
                         {{-- Actions --}}
-                        <div class="items-center justify-between hidden group-hover:flex">
-                            {{-- Edit Button --}}
-                            <a href="{{ route('pages.admin.course.edit', $course->id) }}">
-                                <div
-                                    class="flex items-center justify-center w-8 h-8 mr-2 bg-blue-100 rounded-full ring-2 ring-blue-300 dark:ring-blue-500 hover:bg-blue-200">
-                                    <i class="text-blue-500 fa-solid fa-pen-to-square"></i>
-                                </div>
-                            </a>
+                        @if (request()->routeIs('pages.admin.course'))
+                            <div class="items-center justify-between hidden group-hover:flex">
+                                {{-- Edit Button --}}
+                                <a href="{{ route('pages.admin.course.edit', $course->id) }}">
+                                    <div
+                                        class="flex items-center justify-center w-8 h-8 mr-2 bg-blue-100 rounded-full ring-2 ring-blue-300 dark:ring-blue-500 hover:bg-blue-200">
+                                        <i class="text-blue-500 fa-solid fa-pen-to-square"></i>
+                                    </div>
+                                </a>
 
-                            {{-- Delete Button --}}
-                            <button wire:confirm="Are you sure you want to delete this course?"
-                                wire:click="deleteACourse({{ $course->id }})"
-                                class="flex items-center justify-center w-8 h-8 mr-2 bg-red-100 rounded-full ring-2 ring-red-300 dark:ring-red-500 hover:bg-red-200">
-                                <i class="text-red-500 fa-regular fa-trash-can"></i>
-                            </button>
-                        </div>
+                                {{-- Delete Button --}}
+                                <button wire:confirm="Are you sure you want to delete this course?"
+                                    wire:click="deleteACourse({{ $course->id }})"
+                                    class="flex items-center justify-center w-8 h-8 mr-2 bg-red-100 rounded-full ring-2 ring-red-300 dark:ring-red-500 hover:bg-red-200">
+                                    <i class="text-red-500 fa-regular fa-trash-can"></i>
+                                </button>
+                            </div>
+                        @endif
                     </div>
 
                 </div>
@@ -47,13 +49,15 @@
                     <div class="p-5 cursor-pointer hover:bg-gray-100 dark:hover:bg-gray-900">
                         {{-- Course Name --}}
                         <h5
-                            class="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white line-clamp-2 group-hover:line-clamp-none group-hover:h-min">
+                            class="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white line-clamp-1 group-hover:line-clamp-none group-hover:h-min">
                             {{ $course->course_name }}</h5>
+
                         {{-- Course Short Description --}}
                         <p
-                            class="mb-3 font-normal text-gray-700  dark:text-gray-400 mt-4 h-[100px] line-clamp-4 group-hover:line-clamp-none group-hover:h-min-h-[100px]">
+                            class="mb-3 font-normal text-gray-700  dark:text-gray-400 mt-4 min-h-[100px] line-clamp-4 group-hover:line-clamp-none group-hover:h-min-h-[100px]">
                             {{ $course->course_description }}
                         </p>
+
                         <div class="flex items-center justify-between">
                             <div><i class="fa-solid fa-clock-rotate-left"></i><span class="text-sm">
                                     {{ $course->course_duration }} hrs. read</span>
